야 게시글 작성할 때 id 이거 저장되게해야하는데 안된다?

+++ src/main/java/com/example/demo/dto/BoardDTO.java
@@ -8,7 +8,8 @@ import lombok.Setter;
 @Getter
 @Setter
 public class BoardDTO {
-    private int id;
+
+    private Long id;
     private String title;
     private String content;
     private String filePath;
